+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- Added sellers by cost center feature
+
 ## [1.34.1] - 2023-05-12
 
 ### Fixed